🚀 RELEASE: 2.0 with Multithreading

+++ include/Entry.hpp
@@ -2,33 +2,35 @@
 #define __ENTRY__
 #include <Complex.hpp>
 typedef long long H;
-typedef long long V;
+typedef long double V;
 
 // A class to denote an entry of the map
-class Entry{
+class Entry
+{
 public:
     H h; // Hashed key
     V v; // Value
 
     /* Constructors and Destructors */
     // Constructor with given hash and value
-    Entry(H,V);
+    Entry(H, V);
     // Copy Constructor
-    Entry(const Entry&);
+    Entry(const Entry &);
     // Default destructor
     ~Entry();
 
     /* Comparison operators */
-    bool operator<(const Entry&);
-    bool operator<=(const Entry&);
-    bool operator==(const Entry&);
-    bool operator!=(const Entry&);
-    bool operator>=(const Entry&);
-    bool operator>(const Entry&);
+    bool operator<(const Entry &);
+    bool operator<=(const Entry &);
+    bool operator==(const Entry &);
+    bool operator!=(const Entry &);
+    bool operator>=(const Entry &);
+    bool operator>(const Entry &);
 
     // Pretty printer
-    friend std::ostream& operator<<(std::ostream &cout, Entry &e){
-        cout<<"{ "<<e.h<<" : "<<e.v<<" }";
+    friend std::ostream &operator<<(std::ostream &cout, Entry &e)
+    {
+        cout << "{ " << e.h << " : " << e.v << " }";
         return cout;
     }
 };
